Move check_status_code to commands/upload; add corresponding test
